requirements.txt
langchain==1.2.10
langchain-google-genai==4.2.1
langchain-community==0.4.1
langchain-chroma==1.1.0
sentence-transformers==5.2.3
pdfplumber==0.11.9
chromadb==1.5.1
python-dotenv==1.2.1
